+++ python/fast_mlsirm/__init__.py
@@ -27,6 +27,7 @@ from .serving import (bank_information as bank_information,
                       load_serving_bundle as load_serving_bundle,
                       score_respondents as score_respondents)
 from .preprocessing import irtree_expand as irtree_expand
+from .polytomous import fit_polytomous as fit_polytomous, PolytomousFit as PolytomousFit
 from .simulation import simulate as simulate
 from .test_design import assemble_test_form as assemble_test_form, item_information as item_information, select_cat_item as select_cat_item
 from .types import DimensionalityDiagnostics as DimensionalityDiagnostics, FitDiagnostics as FitDiagnostics, FitResult as FitResult, MLSIRMParams as MLSIRMParams, RecoveryReport as RecoveryReport, SimulationData as SimulationData
@@ -71,6 +72,8 @@ __all__ = [
     "IrtLinkResult",
     "export_serving_bundle",
     "fit",
+    "fit_polytomous",
+    "PolytomousFit",
     "fit_diagnostics",
     "infit_outfit",
     "load_serving_bundle",